Expose exponentiation operator via the API
In contrast to the other binary operators (addition, subtraction,
division, etc.), the exponentation operator is missing in the Clingo
API. That is problematic when power operators are used in an input
program, as a result of which the binary operator field has a value
outside of the declared API enum values, which might then be ignored in
switch statements without any hint from the compiler.

This adds the missing exponentiation operator to the C API as well as
the C++ and Python bindings.