improve parsing of disjunctions
The parser now accepts disjunctions with empty conditions - e.g., `a:.`
Note that for technical reasons the `|` symbol cannot be used after an
empty condition - e.g., `a:;b.` is accepted while `a:|b.` is not.

This also slightly changes the way disjunctions are ordered - e.g.,
`a | b | c.` is no longer grounded as `b | c | a.` but as `a | b | c.`.

This commit partly addresses #171.