improve AST to string conversion
Since there is no direct AST representation for terms of form `f(a;b)`,
they are represented as `f(a);f(b)`. This commit changes printing, so
that such pools are printed in the shorter form. This way, the string
representation of rules containing pools in arguments of predicates is
parsable.

Addresses #171.