make the propagator work for multi-shot solving
This commits enables the propagator to perform (threaded) multi-shot
solving. To test this, there is an option to minimize a csp variable the
same way as clingo-dl does. Furthermore, the prototype can read
clingo-dl instances now.

Finally, it should be possible to use the propagator on realistic domain
sizes now if the sortedcontainers packages is installed. If this is the
case, order literals will be maintained in a SortedDict, which speeds up
propgation if only few order literals have to be introduced.