Add order literals directly for singleton constraints. (#19)
Directly represent singleton constraints by order literals. It is also possible to unroll constraints that can be represented with few order literals but it would be better to do this after the initial propagation when there is more information about upper and lower bounds. Such a feature could supersede what is implemented here.