refactor variable handling
Variables are now represented as integers in constraints and states;
the propagator stores a mapping from symbols to integers. This makes it
possible to introduce unnamed variables while translating.

It also might make sense to do the variable mapping already during
parsing removing it from the builder. This is not yet addressed in this
commit.