progress on stub file
I had to rename clingo.Tuple to clingo.Tuple_ because the name clashed
with the name of the type annotation. Since the name Tuple is so
general, renaming is probably the better strategy.

The main thing that is still missing is to generate the stubs for the
clingo.ast module to get rid of the last error messages. Even replacing
those with typing.Any might work.

Otherwise, some minor error message with mypy are left, which can be
fixed by updating the type signatures.