generate stub files from docstrings (#212)
This commit refines type annotations in the clingo module and adds a
script to generate stub files from the docstrings. The stub files can
be used with systems like mypy. This commit also hides the Slice class
be calling it _Slice and renames Tuple to Tuple_ so that it does not
clash with the typing.Tuple.