fix typing of Applicaton, Observer, Propagator
Unfortunately, it is not (easily) possible to add Python like ABCs to
the clingo interface. As a workaround, these are specified as Protocols
now. Furthermore, because the Python developers decided not to allow for
optional members (see PEP544), all members with default implementations
are omitted.

Future clingo versions might completely avoid such problems by
approaching the implementation of the whole module/package differently.
For example by using the cffi package for the basic interface and
additional Python code to provide a nice high-level interface.