type hints for Application, Observer, Propagator
This commit implements a workaround for the issues mentioned in the
previous commit. It adds empty Python classes created with

    type(name, bases, dict)

to the clingo module. The types for these objects are then provided by
the stub files. The user just has to derive from these classes and the
static type checker can then verify that things fit together.