rework hash_combine and hash_mix for performance
This ended up being a small change, tweaking the hash_combine function
to work more like the boost version and selecting hash_mix based on size
of size_t rather than the argument.  The end result is that for spack
inputs 5.6.2 performs almost exactly as well as 5.5.2, which is a 30%+
speedup in some tests.