fix matchers + weaken comparisons of Ground::Literal
It is important to not provide value-based comparison operators in
Ground::Literal for literals that track state relevant for output.
Literals that have been skipped due to equality would not have this
state. Instead, such duplicated literals could already be dropped in the
input program. Even though, such programs are a bit strange in the first
place.