correctly handle exceptions in C++ api
The current exception design is still flawed because it does not permit
propagating exceptions between bindings, which is relevant when
registering C++ propagators with python.

The only way to achieve something like this is by adding some more
elaborate exception mechanism to the API:
- need some way to set exceptions in an asynchronous context
- need some way to set thread local exceptions

Maybe it is possible to do this using just one function:
- set_error(code, message)
  - set the error in the current context
    - thread local if not in an asynchronous callback
    - control local if in an asynchronous context
      (propagators/on_model/etc.)
- get_error()
  - a previously set error can be restored from the code/message
- clear_error()

The above should make the C++ and Python api much simpler because all
exception pointer management is done in the C api.